+++ prometheus/collectors/go_collector_latest_test.go
@@ -30,7 +30,7 @@ import (
 )
 
 var baseMetrics = []string{
-	"go.gc_duration_seconds",
+	"go_gc_duration_seconds",
 	"go_goroutines",
 	"go_info",
 	"go_memstats_last_gc_time_seconds",